pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adk-channels"
version = "0.1.0"
description = "Multi-channel messaging integration for Google ADK — Slack, Telegram, Webhooks"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
  { name = "OpenCode", email = "agent@opencode.ai" }
]
keywords = ["adk", "google-adk", "slack", "telegram", "webhook", "agent", "chatbot"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Communications :: Chat",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "google-adk>=0.1.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
slack = ["slack-bolt>=1.20.0", "slack-sdk>=3.30.0"]
telegram = ["python-telegram-bot>=21.0"]
webhook = ["fastapi>=0.110.0", "uvicorn>=0.29.0"]
all = [
  "slack-bolt>=1.20.0",
  "slack-sdk>=3.30.0",
  "python-telegram-bot>=21.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
]

[project.scripts]
adk-channels-slack = "examples.slack_fastapi:main"

[project.urls]
Homepage = "https://github.com/opencode/adk-channels"
Repository = "https://github.com/opencode/adk-channels"
Issues = "https://github.com/opencode/adk-channels/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/adk_channels"]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.4.0",
  "mypy>=1.9.0",
  "types-requests",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
